do_claim/is_clan: Fixed a weird NULL CHAR crash bug
Jaya crashed us by using do_claim in Cannaregio. On debug it showed
that fch for notification was NULL. fch = d->character d->character
is set in a loop on descriptor_list. I did not spend enough time in
the debugger to get a full scope of why there was a null ch on a linked
descriptor, but checking that fch != null certainly works.
Also made is_clan do some sanity checking itself.